updateBackup

+++ keyboards/handwired/dactyl_manuform/5x6/keymaps/retro/keymap.c
@@ -179,7 +179,7 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
         case R_THUMB:
             if (record->event.pressed) {
                 layer_on(_R_THUMB);
-                update_tri_layer(_R_THUMB, _L_THUMB, (int) _THUMBS);
+                update_tri_layer(_R_THUMB, _L_THUMB, _THUMBS);
             } else {
                 layer_off(_R_THUMB);
                 update_tri_layer(_R_THUMB, _L_THUMB, (int) _THUMBS);
